fix: Make image pop up successfully
