+++ .idea/misc.xml
@@ -7,10 +7,12 @@
         <entry key="..\:/Users/82105/Desktop/mc firstapp/madcamp-simple-app/app/src/main/res/layout/activity_main.xml" value="0.109375" />
         <entry key="..\:/Users/82105/Desktop/mc firstapp/madcamp-simple-app/app/src/main/res/layout/contacts_list_item.xml" value="0.109375" />
         <entry key="..\:/Users/82105/Desktop/mc firstapp/madcamp-simple-app/app/src/main/res/layout/fragment_contact.xml" value="0.22826086956521738" />
+        <entry key="..\:/Users/82105/Desktop/mc firstapp/madcamp-simple-app/app/src/main/res/layout/fragment_gallery.xml" value="0.152" />
+        <entry key="..\:/Users/82105/Desktop/mc firstapp/madcamp-simple-app/app/src/main/res/layout/fragment_gallery_click_image.xml" value="0.14895833333333333" />
         <entry key="..\:/Users/82105/Desktop/mc firstapp/madcamp-simple-app/app/src/main/res/layout/fragment_image.xml" value="0.109375" />
         <entry key="..\:/Users/82105/Desktop/mc firstapp/madcamp-simple-app/app/src/main/res/layout/fragment_phone.xml" value="0.109375" />
         <entry key="..\:/Users/82105/Desktop/mc firstapp/madcamp-simple-app/app/src/main/res/layout/fragment_todo.xml" value="0.264" />
-        <entry key="..\:/Users/82105/Desktop/mc firstapp/madcamp-simple-app/app/src/main/res/layout/gallery_item.xml" value="0.24947916666666667" />
+        <entry key="..\:/Users/82105/Desktop/mc firstapp/madcamp-simple-app/app/src/main/res/layout/gallery_item.xml" value="0.152" />
         <entry key="..\:/Users/82105/Desktop/mc firstapp/madcamp-simple-app/app/src/main/res/layout/todo_item.xml" value="0.24947916666666667" />
         <entry key="..\:/Users/82105/Desktop/mc firstapp/madcamp-simple-app/app/src/main/res/menu/main_3tabs.xml" value="0.2" />
       </map>
