+++ app/src/main/java/com/example/themostusefulapp/Gallery/GalleryAdapter.java
@@ -1,20 +1,16 @@
-package com.example.themostusefulapp;
+package com.example.themostusefulapp.Gallery;
 
-import android.Manifest;
 import android.content.Context;
-import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.graphics.BitmapFactory;
-import android.net.Uri;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
-import androidx.core.app.ActivityCompat;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.themostusefulapp.R;
 
 import java.util.ArrayList;
 
@@ -23,6 +19,17 @@ public class GalleryAdapter extends RecyclerView.Adapter<GalleryAdapter.ViewHold
     private final ArrayList<GalleryData> gallery = new ArrayList<>();
     private final Context context;
 
+//    //커스텀 리스너 인터페이스 정의
+//    public interface OnItemClickListener{
+//        void onItemClick(View v, int pos);
+//    }
+//
+//    //리스너 객체를 전달하는 메소드 setOnItemClickListener와 전달된 객체를 저장할 변수 mListener 추가
+//    private  OnItemClickListener mListener=null;
+//    public void setOnItemClickListener(OnItemClickListener listener){
+//        this.mListener=listener;
+//    }
+
     public GalleryAdapter(Context context) {
         this.context = context;
     }
@@ -30,29 +37,16 @@ public class GalleryAdapter extends RecyclerView.Adapter<GalleryAdapter.ViewHold
     @NonNull
     @Override
     public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater
-                .from(parent.getContext())
-                .inflate(
-                        R.layout.gallery_item,
-                        parent,
-                        false
-                );
+        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.gallery_item, parent, false);
         return new GalleryAdapter.ViewHolder(view);
     }
 
     @Override
     public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
-        //Contact 부분
-        GalleryData galleryData = gallery.get(position);
+        GalleryData galleryData=gallery.get(position);
 
         holder.imageView.setImageBitmap(BitmapFactory.decodeStream(galleryData.getImageView()));
 
-        holder.imageView.setOnClickListener(view -> {
-            if (ActivityCompat.checkSelfPermission(context, Manifest.permission.READ_EXTERNAL_STORAGE) == PackageManager.PERMISSION_DENIED) {
-                Toast.makeText(context, "Phone permission denied", Toast.LENGTH_SHORT).show();
-                return;
-            }
-        });
     }
 
     @Override
@@ -60,7 +54,7 @@ public class GalleryAdapter extends RecyclerView.Adapter<GalleryAdapter.ViewHold
         return gallery.size();
     }
 
-    public void addContact(GalleryData galleryData) {
+    public void addImage(GalleryData galleryData) {
         this.gallery.add(galleryData);
         notifyDataSetChanged();
     }
@@ -70,7 +64,16 @@ public class GalleryAdapter extends RecyclerView.Adapter<GalleryAdapter.ViewHold
 
         public ViewHolder(@NonNull View itemView) {
             super(itemView);
-            imageView = itemView.findViewById(R.id.image);
+            imageView=itemView.findViewById(R.id.image);
+//            //아이템 클릭 이벤트 핸들러 메소드에서 리스너 객체 메소드(onItemClick) 호출
+//            itemView.setOnClickListener(v -> {
+//                int pos=getAdapterPosition();
+//                if(pos!=RecyclerView.NO_POSITION){
+//                    if(mListener!=null){
+//                        mListener.onItemClick(v, pos);
+//                    }
+//                }
+//            });
 
         }
     }
