+++ app/src/main/java/com/example/themostusefulapp/Gallery/GalleryClickImage.java
@@ -1,0 +1,79 @@
+package com.example.themostusefulapp.Gallery;
+
+import android.content.Intent;
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+
+import com.example.themostusefulapp.ThreeMenuFragment.FragmentGallery;
+import com.example.themostusefulapp.R;
+
+/**
+ * A simple {@link Fragment} subclass.
+ * Use the {@link GalleryClickImage#newInstance} factory method to
+ * create an instance of this fragment.
+ */
+public class GalleryClickImage extends Fragment {
+
+    // TODO: Rename parameter arguments, choose names that match
+    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
+    private static final String ARG_PARAM1 = "param1";
+    private static final String ARG_PARAM2 = "param2";
+
+    // TODO: Rename and change types of parameters
+    private String mParam1;
+    private String mParam2;
+
+    public GalleryClickImage() {
+        // Required empty public constructor
+    }
+
+    /**
+     * Use this factory method to create a new instance of
+     * this fragment using the provided parameters.
+     *
+     * @param param1 Parameter 1.
+     * @param param2 Parameter 2.
+     * @return A new instance of fragment GalleryClickImage.
+     */
+    // TODO: Rename and change types and number of parameters
+    public static GalleryClickImage newInstance(String param1, String param2) {
+        GalleryClickImage fragment = new GalleryClickImage();
+        Bundle args = new Bundle();
+        args.putString(ARG_PARAM1, param1);
+        args.putString(ARG_PARAM2, param2);
+        fragment.setArguments(args);
+        return fragment;
+    }
+
+    Button btn_back;
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        if (getArguments() != null) {
+            mParam1 = getArguments().getString(ARG_PARAM1);
+            mParam2 = getArguments().getString(ARG_PARAM2);
+        }
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        btn_back=getView().findViewById(R.id.btn_move);
+        btn_back.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(getActivity(), FragmentGallery.class);
+                startActivity(intent);
+            }
+        });
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_gallery_click_image, container, false);
+    }
+}
