+++ app/src/main/java/com/example/themostusefulapp/ThreeMenuFragment/FragmentContacts.java
@@ -1,4 +1,4 @@
-package com.example.themostusefulapp;
+package com.example.themostusefulapp.ThreeMenuFragment;
 
 import android.Manifest;
 import android.content.ContentUris;
@@ -8,12 +8,10 @@ import android.database.Cursor;
 import android.net.Uri;
 import android.os.Bundle;
 import android.provider.ContactsContract;
-import android.provider.MediaStore;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ImageView;
 import android.widget.Toast;
 
 import androidx.activity.result.ActivityResultLauncher;
@@ -25,18 +23,22 @@ import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.themostusefulapp.Contact.Contact;
+import com.example.themostusefulapp.Contact.ContactsRecViewAdapter;
+import com.example.themostusefulapp.R;
+
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.Map;
 
-public class FragmentGallery extends Fragment {
 
+public class FragmentContacts extends Fragment {
     private ActivityResultLauncher<String[]> requestPermissionLauncher;
-    private GalleryAdapter adapter;
+    private ContactsRecViewAdapter adapter;
 
-    public FragmentGallery() {
-        super(R.layout.fragment_gallery);
+    public FragmentContacts() {
+        super(R.layout.fragment_contacts);
     }
 
     @Override
@@ -57,9 +59,9 @@ public class FragmentGallery extends Fragment {
                         if (permission.getValue()) {
                             // Permission is granted. Continue the action or workflow in your app.
                             Log.d("PERMISSION", "requestPermissionLauncher :: " + permission.getKey() + " granted");
-                            if (permission.getKey().equals(Manifest.permission.READ_EXTERNAL_STORAGE)) {
-                                Log.d("PERMISSION", "Load gallery data from the device for the first time");
-                                getGalleryList(context);
+                            if (permission.getKey().equals(Manifest.permission.READ_CONTACTS)) {
+                                Log.d("PERMISSION", "Load contacts data from the device for the first time");
+                                getContactList(context);
                             }
                         } else {
                             // Explain to the user that the feature is unavailable because the
@@ -74,20 +76,38 @@ public class FragmentGallery extends Fragment {
 
         /* ==== create RecyclerView ==== */
 
-        adapter = new GalleryAdapter(context);
-        RecyclerView galleryRecycleView = view.findViewById(R.id.galleryRecycleView);
-        galleryRecycleView.setAdapter(adapter);
-        galleryRecycleView.setLayoutManager(new LinearLayoutManager(context));
+        adapter = new ContactsRecViewAdapter(context);
+        RecyclerView contactsRecView = view.findViewById(R.id.contactsRecView);
+        contactsRecView.setAdapter(adapter);
+        contactsRecView.setLayoutManager(new LinearLayoutManager(context));
 
         /* ==== create new contact ==== */
 
-        ImageView imageView = view.findViewById(R.id.image);
+        EditText editName = view.findViewById(R.id.editName);
+        EditText editNumber = view.findViewById(R.id.editNumber);
+        Button contactAddBtn = view.findViewById(R.id.contactAddBtn);
+
+        // TODO: Get photo from user
+        contactAddBtn.setOnClickListener(v -> {
+            String name = editName.getText().toString();
+            String number = editNumber.getText().toString();
+
+            if (name.isEmpty()) {
+                Toast.makeText(context, "Name is empty", Toast.LENGTH_SHORT).show();
+                return;
+            }
+            if (number.isEmpty()) {
+                Toast.makeText(context, "Number is empty", Toast.LENGTH_SHORT).show();
+                return;
+            }
+
+            adapter.addContact(new Contact(name, number, null));
+        });
 
         /* ==== request contacts permission ==== */
 
-        requestPermission(context, new String[]{Manifest.permission.READ_EXTERNAL_STORAGE});
+        requestPermission(context, new String[]{Manifest.permission.READ_CONTACTS, Manifest.permission.CALL_PHONE});
     }
-
 
     private void requestPermission(Context context, String[] permissions) {
         final ArrayList<String> permissionsToRequest = new ArrayList<>();
@@ -96,9 +116,9 @@ public class FragmentGallery extends Fragment {
             // You can use the API that requires the permission.
             if (ContextCompat.checkSelfPermission(context, permission) == PackageManager.PERMISSION_GRANTED) {
                 Log.d("PERMISSION", "Permission " + permission + " already granted");
-                if (permission.equals(Manifest.permission.READ_EXTERNAL_STORAGE)) {
-                    Log.d("PERMISSION", "Load gallery data from the device");
-                    getGalleryList(context);
+                if (permission.equals(Manifest.permission.READ_CONTACTS)) {
+                    Log.d("PERMISSION", "Load contacts data from the device");
+                    getContactList(context);
                 }
                 continue;
             }
@@ -121,50 +141,50 @@ public class FragmentGallery extends Fragment {
             requestPermissionLauncher.launch(permissionsToRequest.toArray(new String[0]));
     }
 
-    private void getGalleryList(Context context) {
+    private void getContactList(Context context) {
         Cursor cursor = context.getContentResolver().query(
-                MediaStore.Images.Contacts.CONTENT_URI,
+                ContactsContract.Contacts.CONTENT_URI,
                 null,
                 null,
                 null,
-                MediaStore.Images.CommonDataKinds.Phone.DISPLAY_NAME + " ASC "
+                ContactsContract.CommonDataKinds.Phone.DISPLAY_NAME + " ASC "
         );
 
         while (cursor.moveToNext()) {
             String id = cursor.getString(cursor.getColumnIndexOrThrow(
-                    Images.Contacts._ID
+                    ContactsContract.Contacts._ID
             ));
             String name = cursor.getString(cursor.getColumnIndexOrThrow(
-                    Images.Contacts.DISPLAY_NAME
+                    ContactsContract.Contacts.DISPLAY_NAME
             ));
 
-            Cursor galleryCursor = context.getContentResolver().query(
-                    Images.CommonDataKinds.Phone.CONTENT_URI,
+            Cursor phoneCursor = context.getContentResolver().query(
+                    ContactsContract.CommonDataKinds.Phone.CONTENT_URI,
                     null,
-                    Images.CommonDataKinds.Phone.CONTACT_ID + "=?",
+                    ContactsContract.CommonDataKinds.Phone.CONTACT_ID + "=?",
                     new String[]{id},
                     null
             );
 
-            if (galleryCursor.moveToNext()) {
-                String number = galleryCursor.getString(galleryCursor.getColumnIndexOrThrow(
-                        Images.Phone.NUMBER
+            if (phoneCursor.moveToNext()) {
+                String number = phoneCursor.getString(phoneCursor.getColumnIndexOrThrow(
+                        ContactsContract.CommonDataKinds.Phone.NUMBER
                 ));
-                InputStream gallery = getGalleryStream(context, Long.parseLong(id));
+                InputStream photo = getPhotoStream(context, Long.parseLong(id));
 
-                adapter.addContact(new GalleryData(gallery));
+                adapter.addContact(new Contact(name, number, photo));
             }
-            galleryCursor.close();
+            phoneCursor.close();
         }
         cursor.close();
     }
 
-    private InputStream getGalleryStream(Context context, long contactId) {
-        Uri contactUri = ContentUris.withAppendedId(Images.CONTENT_URI, contactId);
-        Uri photoUri = Uri.withAppendedPath(contactUri, Images.Photo.CONTENT_DIRECTORY);
+    private InputStream getPhotoStream(Context context, long contactId) {
+        Uri contactUri = ContentUris.withAppendedId(ContactsContract.Contacts.CONTENT_URI, contactId);
+        Uri photoUri = Uri.withAppendedPath(contactUri, ContactsContract.Contacts.Photo.CONTENT_DIRECTORY);
         Cursor cursor = context.getContentResolver().query(
                 photoUri,
-                new String[]{Images.Photo.PHOTO},
+                new String[]{ContactsContract.Contacts.Photo.PHOTO},
                 null,
                 null,
                 null
