+++ app/src/main/java/com/example/themostusefulapp/ThreeMenuFragment/FragmentTodo.java
@@ -1,4 +1,4 @@
-package com.example.themostusefulapp;
+package com.example.themostusefulapp.ThreeMenuFragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -6,6 +6,8 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
+
+import com.example.themostusefulapp.R;
 
 public class FragmentTodo extends Fragment {
     public FragmentTodo() {
