+++ app/src/main/res/layout/gallery_item.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<GridLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
+<androidx.cardview.widget.CardView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content">
 
     <ImageView
         android:id="@+id/image"
@@ -10,4 +11,4 @@
         android:src="@drawable/ic_image_solid"
         />
 
-</GridLayout>
+</androidx.cardview.widget.CardView>
